Create scenarios for Bogota

Creates three scenarios where, in each one, the mode share is elevated by a set percentage. The scenario-modes are cycle, car, and bus.

create_bogota_scenarios(trip_set)

Arguments

trip_set

data frame, baseline scenario trips

Value

list of baseline scenario and three scenarios

Details

This function creates three scenarios increasing the mode shares of cycling, car and bus by the same pre-defined percentage. We assume that:

- the total number of trips remains the same but trips from other modes (apart from truck, bus driver, car driver and commercial motorcycle trips which remain unchanged initially) are converted to the mode in question; truck and commercial motorcycle trips remain constant across all scenarios whilst bus driver and car driver trips are updated based on the new amount of car and bus trips once the increase in mode share has been conducted for each scenario - the percentage share across the three distance bands of the mode that is increased is preserved, i.e. if 10 then after increasing the cycling mode share by x have 10 - For each scenario we always convert the same in question, independently of the original overall mode share of that mode. - we preserve the proportion of trips in each distance band. E.g. if 20 all trips are in distance band 0-2km, then for each scenario 20 trips are still in distance band 0-2km.

Example:

Assume that there are only two distance bands A and B and that 80 cycling trips lie in distance band A and the remaining 20 band B. Assume that 60 band B. Assume we want to increase the cycling mode share by 5

Then, we need to convert 5 to cycling trips and 5 distance band B. Overall, this leads to an increase of (5 of cycling trips, whilst preserving the cycling mode shares of 80 band A and 20 distance band A and 40 preserved,