Creates synthetic population

Creates a synthetic population by matching individuals in the trip set to individuals in the PA set

create_synth_pop(raw_trip_set)

Arguments

raw_trip_set

data frame of raw trips taken, bus_driver, new motorcycle and truck trips have already been added

Value

the synthetic population and the trip set which has been pruned

Details

The function performs the following steps:

- adds age category to trip and physical activity datasets - match people in trip data with people in the physical activity dataset - create a synthetic population by taking the unique participant ids together with age and gender information from the trip data (not including bus driver, truck, car driver and commercial motorcycle trips) - to assign non - travel (i.e. work and leisure) physical activity MMET values to this synthetic population, the following steps are performed: - for each sex and age category, find the proportion of people with zero work and leisure MMET values and also find the list of people with non-zero MMET values - if BACKGROUND_PA_CONFIDENCE < 1 when calling the value of information script, i.e. when input values are sampled from distributions, a beta distribution is built from which the proportion of people with zero work and leisure MMET values is sample using the the known proportion as mean of that distribution - sample with replacement from 0 MMET values and the list of people with non-zero MMET values using the proportion of people with zero work and leisure MMET values and assign those sampled MMET values to the synthetic population - remove any non-travel trips from the trip data