Cascade of computations that form ITHIM-Global

Ordered set of computations as part of ITHIM-Global that calculates the required output parameters

Usage

ithim_calculation_sequence(ithim_object, seed = 1)

Arguments

ithim_object

name of disease

seed

Value

ithim_object - list of items making up the ithim result

Details

This function performs the following steps:

extract all lists and variables from the ithim_object list

air pollution pathway:

calculate the PM2.5 exposure for each person in the synthetic population and PM2.5 emissions for each mode and scenario (scenario_pm_calculations())

calculate the CO2 emissions for each mode and scenario (scenario_co2_calculations())

assign relative risk to each person in the synthetic population for each disease related to PM pollution and each scenario based on the individual PM exposure levels (gen_ap_rr())

physical activity pathway:

calculate total mMETs for each person in the synthetic population (total_mmet())

assign relative risk to each person in the synthetic population for each disease related to physical activity levels and each scenario based on the individual mMET values (gen_pa_rr())

physical activity and air pollution combined:

combine the PA and AP datasets by joining the two datasets. For disease affected by both PA and AP calculate the joined relative risk by multiplying the PA and AP relative risks (combined_rr_ap_pa())

calculate the health burden (Yll and deaths) for each disease and age and sex category (health_burden()):

calculate the health burden (Yll and deaths) for each disease and age and sex category. Combine the AP and PA pathways for diseases affected by both AP and PA

if running in constant mode also calculate the health burden for both the AP and PA pathways separately

injury pathway:

estimate the injury deaths for the baseline and each scenario by age and sex category, also estimate the total injury deaths counts for the who-hit-whom and no-other-vehicle matrices by casualty (and strike) mode again for the baseline and each scenario (injuries_function2())

if running in constant mode include upper and lower confidence intervals

calculate the years of life lost from the injury deaths (injury_death_to_yll())

combine all pathways using the outputs from 3. and 4.:

combine the AP, PA and injury health burden data for ylls and deaths (join_hb_and_injury()) for all diseases, injuries and scenarios