Calculate relative risk given PM exposure level

Calculate the relative risk (RR) for each person in the synthetic population for each disease related to air pollution and each scenario based on the individual PM exposure levels

Usage

AP_dose_response(cause, dose, quantile, confidence_intervals = F)

Arguments

cause

name of disease

dose

vector of PM exposure levels from individuals for a given age range and scenario

quantile

quantile of the dose response functions to be used as output value

confidence_intervals

logic: whether or not to return confidence intervals

Value

data frame of relative risks for each person in the synthetic population and each AP related disease in each scenario

Details

This function performs the following steps:

various checks to ensure the correct PM exposure levels (doses) and disease are read in

get the lookup table for the required dose response functions which contains for each dose the median RR and the upper and lower confidence interval RR values

for the PM exposure doses in the synthetic population find the needed RR by extrapolating the dose responses given in the lookup table

if a confidence interval is required, repeat this interpolation using the upper and lower confidence RR values

if the required quantile is not 0.5, i.e. the median, then find the required RR value by defining a normal function with mean the median RR value for that dose and standard deviations defined by the upper and lower confidence RR values. Use the correct quantile from this normal function as your RR value.