Personal motorcycle trips for Chilean cities

Function to create personal motorcycle trips for some Chilean cities

add_motorcycle_trips_Chile(raw_trip_set, PROPORTION_MOTORCYCLE_TRIPS)

Arguments

raw_trip_set

data frame of trips from travel survey

PROPORTION_MOTORCYCLE_TRIPS

proportion of trips in travel survey that are to be added as personal motorcycle trips

Value

original trip data with additional motorcycle trips added

Details

This function is used to create personal motorcycle trips for those Chilean cities which do not have any motorcycle trips in their travel surveys. It uses as input the existing travel survey data for that city and the proportion of these travel survey trips that are to be added as additional new motorcycle trips.

Based on an analysis of the motorcycle trips in the travel surveys of Santiago, San Antonio, Valparaiso and Puerto Montt it makes assumptions about the split between male and female motorcyclists, the number of trips per person and the truncated normal distributions of the distances and ages of the motorcycle trips to be added. It creates new trips by assuming that each such trip only has one stage, i.e. that the trip duration and distance equals the stage duration and distance.