Routine to sample or set parameters for ITHIM

Parameters have two options: to be set to a constant, and to be sampled from a prespecified distribution. Each parameter is given as an argument of length 1 or 2. If length 1, it's constant, and set to the global environment. If length 2, a distribution is defined and sampled from NSAMPLE times. There are some exceptions, listed below.

ithim_setup_parameters(NSAMPLES = 1, BUS_WALK_TIME = 5,
  MMET_CYCLING = 4.63, MMET_WALKING = 2.53, PM_CONC_BASE = 50,
  PM_TRANS_SHARE = 0.225, PA_DOSE_RESPONSE_QUANTILE = F,
  AP_DOSE_RESPONSE_QUANTILE = F, BACKGROUND_PA_SCALAR = 1,
  BACKGROUND_PA_CONFIDENCE = 1, INJURY_REPORTING_RATE = 1,
  CHRONIC_DISEASE_SCALAR = 1, DAY_TO_WEEK_TRAVEL_SCALAR = 7,
  INJURY_LINEARITY = 1, CASUALTY_EXPONENT_FRACTION = 0.5,
  BUS_TO_PASSENGER_RATIO = 0.022, TRUCK_TO_CAR_RATIO = 0.21,
  EMISSION_INVENTORY_CONFIDENCE = 1, DISTANCE_SCALAR_CAR_TAXI = 1,
  DISTANCE_SCALAR_WALKING = 1, DISTANCE_SCALAR_PT = 1,
  DISTANCE_SCALAR_CYCLING = 1, DISTANCE_SCALAR_MOTORCYCLE = 1)

Arguments

NSAMPLES
constant integer: number of samples to take

BUS_WALK_TIME
lognormal parameter: duration of walk to PT

MMET_CYCLING
lognormal parameter: mMETs when cycling

MMET_WALKING
lognormal parameter: mMETs when walking

PM_CONC_BASE
lognormal parameter: background PM2.5 concentration

PM_TRANS_SHARE
beta parameter: fraction of background PM2.5 attributable to transport

PA_DOSE_RESPONSE_QUANTILE
logic: whether or not to sample from PA RR DR functions

AP_DOSE_RESPONSE_QUANTILE
logic: whether or not to sample from AP RR DR functions

BACKGROUND_PA_SCALAR
lognormal parameter: reporting scalar for PA

BACKGROUND_PA_CONFIDENCE
beta parameter: confidence in accuracy of PA survey

INJURY_REPORTING_RATE
lognormal parameter: rate of injury reporting

CHRONIC_DISEASE_SCALAR
lognormal parameter: scalar for background disease rates

DAY_TO_WEEK_TRAVEL_SCALAR
beta parameter: rate of scaling travel from one day to one week

INJURY_LINEARITY
lognormal parameter: linearity of injuries in space

CASUALTY_EXPONENT_FRACTION
beta parameter: casualty contribution to linearity of scalaing of injuries in space

BUS_TO_PASSENGER_RATIO
beta parameter: number of buses per passenger

TRUCK_TO_CAR_RATIO
beta parameter: number of trucks per car

EMISSION_INVENTORY_CONFIDENCE
beta parameter: confidence in accuracy of emission inventory

DISTANCE_SCALAR_CAR_TAXI
lognormal parameter: scalar for car distance travelled

DISTANCE_SCALAR_WALKING
lognormal parameter: scalar for walking distance travelled

DISTANCE_SCALAR_PT
lognormal parameter: scalar for PT distance travelled

DISTANCE_SCALAR_CYCLING
lognormal parameter: scalar for cycling distance travelled

DISTANCE_SCALAR_MOTORCYCLE
lognormal parameter: scalar for motorcycle distance travelled

Value

list of samples of uncertain parameters