Run the set up scripts for ITHIM

Sets up the basic ITHIM object for onward calculation. Data loading, processing and harmonisation. Setting global values.

run_ithim_setup(
  seed = 1,
  CITY = "accra",
  speeds = NULL,
  PM_emission_inventory = NULL,
  CO2_emission_inventory = NULL,
  DIST_CAT = c("0-6 km", "7-9 km", "10+ km"),
  AGE_RANGE = c(0, 150),
  ADD_WALK_TO_PT_TRIPS = T,
  ADD_BUS_DRIVERS = T,
  ADD_CAR_DRIVERS = T,
  ADD_TRUCK_DRIVERS = T,
  ADD_MOTORCYCLE_FLEET = F,
  ADD_PERSONAL_MOTORCYCLE_TRIPS = "no",
  REFERENCE_SCENARIO = "baseline",
  PATH_TO_LOCAL_DATA = NULL,
  NSAMPLES = 1,
  BUS_WALK_TIME = 5,
  RAIL_WALK_TIME = 15,
  MMET_CYCLING = 4.63,
  MMET_WALKING = 2.53,
  PM_CONC_BASE = 50,
  PM_TRANS_SHARE = 0.225,
  PA_DOSE_RESPONSE_QUANTILE = F,
  AP_DOSE_RESPONSE_QUANTILE = F,
  BACKGROUND_PA_SCALAR = 1,
  BACKGROUND_PA_CONFIDENCE = 1,
  INJURY_REPORTING_RATE = 1,
  CHRONIC_DISEASE_SCALAR = 1,
  DAY_TO_WEEK_TRAVEL_SCALAR = 7,
  SIN_EXPONENT_SUM = 2,
  CASUALTY_EXPONENT_FRACTION = 0.5,
  SIN_EXPONENT_SUM_NOV = 1,
  SIN_EXPONENT_SUM_CYCLE = 2,
  CASUALTY_EXPONENT_FRACTION_CYCLE = 0.5,
  SIN_EXPONENT_SUM_PED = 2,
  CASUALTY_EXPONENT_FRACTION_PED = 0.5,
  SIN_EXPONENT_SUM_VEH = 2,
  CASUALTY_EXPONENT_FRACTION_VEH = 0.5,
  CALL_INDIVIDUAL_SIN = F,
  BUS_TO_PASSENGER_RATIO = 0.022,
  CAR_OCCUPANCY_RATIO = 0.6,
  TRUCK_TO_CAR_RATIO = 0.21,
  FLEET_TO_MOTORCYCLE_RATIO = 0.01,
  PROPORTION_MOTORCYCLE_TRIPS = 0,
  PM_EMISSION_INVENTORY_CONFIDENCE = 1,
  CO2_EMISSION_INVENTORY_CONFIDENCE = 1,
  DISTANCE_SCALAR_CAR_TAXI = 1,
  DISTANCE_SCALAR_WALKING = 1,
  DISTANCE_SCALAR_PT = 1,
  DISTANCE_SCALAR_CYCLING = 1,
  DISTANCE_SCALAR_MOTORCYCLE = 1,
  SCENARIO_NAME = "GLOBAL",
  SCENARIO_INCREASE = 0.05,
  BUS_DRIVER_PROP_MALE = 0.99,
  BUS_DRIVER_MALE_AGERANGE = "18, 65",
  BUS_DRIVER_FEMALE_AGERANGE = "18, 65",
  TRUCK_DRIVER_PROP_MALE = 0.99,
  TRUCK_DRIVER_MALE_AGERANGE = "18, 65",
  TRUCK_DRIVER_FEMALE_AGERANGE = "18, 65",
  COMMERCIAL_MBIKE_PROP_MALE = 0.99,
  COMMERCIAL_MBIKE_MALE_AGERANGE = "18, 65",
  COMMERCIAL_MBIKE_FEMALE_AGERANGE = "18, 65",
  MINIMUM_PT_TIME = 3
)

Arguments

seed

random seed

CITY

name of the city, and name of the directory containing city data files

speeds

named list of mode speeds

PM_emission_inventory

named list of mode emissions

CO2_emission_inventory

named list of mode emissions

DIST_CAT

vector string of distance categories in the form '0-6'. (The unit is assumed to be the same as in the trip set.)

AGE_RANGE

vector of minimum and maximum ages to include

ADD_WALK_TO_PT_TRIPS

logic: whether or not to add short walks to all PT trips

ADD_BUS_DRIVERS

logic: whether or not to add bus drivers

ADD_CAR_DRIVERS

logic: whether or not to find and add distance travelled by individual cars, denoted by car drivers

ADD_TRUCK_DRIVERS

logic: whether or not to add truck drivers

ADD_MOTORCYCLE_FLEET

logic: whether or not to add additional commercial motorcycle fleet as ghost trips

ADD_PERSONAL_MOTORCYCLE_TRIPS

character: if 'no' does not add motorcycle trips otherwise set to geographic region which defines the set-up of the motorcycle trips to be added

REFERENCE_SCENARIO

which scenario forms the reference for the health comparison

PATH_TO_LOCAL_DATA

path to CITY directory, if not using package

NSAMPLES

constant integer: number of samples to take

BUS_WALK_TIME

lognormal parameter: duration of walk to Bus

RAIL_WALK_TIME

lognormal parameter: duration of walk to Rail

MMET_CYCLING

lognormal parameter: mMETs when cycling

MMET_WALKING

lognormal parameter: mMETs when walking

PM_CONC_BASE

lognormal parameter: background PM2.5 concentration

PM_TRANS_SHARE

beta parameter: fraction of background PM2.5 attributable to transport

PA_DOSE_RESPONSE_QUANTILE

logic: whether or not to sample from PA RR DR functions

AP_DOSE_RESPONSE_QUANTILE

logic: whether or not to sample from AP RR DR functions

BACKGROUND_PA_SCALAR

lognormal parameter: reporting scalar for PA

BACKGROUND_PA_CONFIDENCE

beta parameter: confidence in accuracy of PA survey

INJURY_REPORTING_RATE

lognormal parameter: rate of injury reporting

CHRONIC_DISEASE_SCALAR

lognormal parameter: scalar for background disease rates

DAY_TO_WEEK_TRAVEL_SCALAR

beta parameter: rate of scaling travel from one day to one week - CURRENTLY used as constant only (using as beta parameter would need some further considerations)

SIN_EXPONENT_SUM

lognormal parameter: linearity of injuries with respect to two modes. SIN_EXPONENT_SUM=2 means no safety in numbers.

CASUALTY_EXPONENT_FRACTION

beta parameter: casualty contribution to SIN_EXPONENT_SUM

SIN_EXPONENT_SUM_NOV

lognormal parameter: linearity of injuries with respect to two modes where strike mode = NOV. SIN_EXPONENT_SUM=2 means no safety in numbers.

SIN_EXPONENT_SUM_CYCLE

lognormal parameter: linearity of injuries with respect to two modes where victim mode = cycle. SIN_EXPONENT_SUM=2 means no safety in numbers.

CASUALTY_EXPONENT_FRACTION_CYCLE

beta parameter: casualty contribution to SIN_EXPONENT_SUM where victim mode = cycle.

SIN_EXPONENT_SUM_PED

lognormal parameter: linearity of injuries with respect to two modes where victim mode = pedestrian. SIN_EXPONENT_SUM=2 means no safety in numbers.

CASUALTY_EXPONENT_FRACTION_PED

beta parameter: casualty contribution to SIN_EXPONENT_SUM where victim mode = pedestrian

SIN_EXPONENT_SUM_VEH

lognormal parameter: linearity of injuries with respect to two modes where victim mode = a vehicle. SIN_EXPONENT_SUM=2 means no safety in numbers.

CASUALTY_EXPONENT_FRACTION_VEH

beta parameter: casualty contribution to SIN_EXPONENT_SUM where victim mode = a vehicle.

CALL_INDIVIDUAL_SIN

logic: whether or not to call the safety in number coefficients for individual vehicles or use the same coefficients for all modes.

BUS_TO_PASSENGER_RATIO

beta parameter: number of buses per passenger

CAR_OCCUPANCY_RATIO

beta parameter: number of people per car (including driver)

TRUCK_TO_CAR_RATIO

beta parameter: number of trucks per car

FLEET_TO_MOTORCYCLE_RATIO

beta parameter: fraction of total motorcycles that's fleet

PROPORTION_MOTORCYCLE_TRIPS

beta parameter: proportion of trips that are to be added as motorcycle trips

PM_EMISSION_INVENTORY_CONFIDENCE

beta parameter: confidence in accuracy of emission inventory

CO2_EMISSION_INVENTORY_CONFIDENCE

beta parameter: confidence in accuracy of emission inventory

DISTANCE_SCALAR_CAR_TAXI

lognormal parameter: scalar for car distance travelled

DISTANCE_SCALAR_WALKING

lognormal parameter: scalar for walking distance travelled

DISTANCE_SCALAR_PT

lognormal parameter: scalar for PT distance travelled

DISTANCE_SCALAR_CYCLING

lognormal parameter: scalar for cycling distance travelled

DISTANCE_SCALAR_MOTORCYCLE

lognormal parameter: scalar for motorcycle distance travelled

SCENARIO_NAME

name of the scenarios (currently supports: TEST_WALK_SCENARIO, TEST_CYCLE_SCENARIO, MAX_MODE_SHARE_SCENARIO, LATAM, GLOBAL, AFRICA_INDIA, BOGOTA)

SCENARIO_INCREASE

increase of given mode in each scenario (currently used in GLOBAL, BOGOTA, LATAM and AFRICA_INDIA scenarios)

BUS_DRIVER_PROP_MALE

scalar parameter: proportion of bus drivers that are male

BUS_DRIVER_MALE_AGERANGE

character parameter: age range of male bus drivers

BUS_DRIVER_FEMALE_AGERANGE

character parameter: age range of female bus drivers

TRUCK_DRIVER_PROP_MALE

scalar parameter: proportion of truck drivers that are male

TRUCK_DRIVER_MALE_AGERANGE

character parameter: age range of male truck drivers

TRUCK_DRIVER_FEMALE_AGERANGE

character parameter: age range of female truck drivers

COMMERCIAL_MBIKE_PROP_MALE

scalar parameter: proportion of commercial motorcycle drivers that are male

COMMERCIAL_MBIKE_MALE_AGERANGE

character parameter: age range of male commercial motorcycle drivers

COMMERCIAL_MBIKE_FEMALE_AGERANGE

character parameter: age range of female commercial motorcycle drivers

MINIMUM_PT_TIME

minimum time that person spends on public transport

setup_call_summary_filename

name to write setup call summary to

Value

ithim_object list of objects for onward use.

Details

Parameters have two options: to be set to a constant, and to be sampled from a pre-specified distribution. Each parameter is given as an argument of length 1 or 2. If length 1, it's constant, and set to the global environment. If length 2, a distribution is defined and sampled from NSAMPLE times. (There are some exceptions)

This function performs the following steps: - check whether valid scenario name is called, get error message if not - set various parameters as global parameters - find the path to the local data - set fixed parameters for air pollution inhalation - define the mode speeds: - set default speeds for the various modes - update the default speeds with city specific mode speeds if these were given as input parameters - ensure similar modes have the same speed assigned - define dataframe with modes and speeds - define PM emissions inventory - set default emission values - update default values if city specific values were given as input parameters - define PM emissions inventory - set default emission values - update default values if city specific values were given as input parameters - load and processes data from files by calling ithim_load_data.R - call ithim_setup_parameters.R to set the given input parameters to the global environment if running in constant mode or to create nsamples samples from the given distributions for each of the input parameters if running in sample mode - set flags which cause certain parts of the model to be called at a later stage (ithim_uncertainty.R) IF certain input parameters were sampled from a distribution - call complete_trip_distance_duration.R to add any missing stage or distance information to TRIP_SET - if none of the corresponding input parameters were sampled from a distribution, call set_vehicle_inventory.R to create a dataframe with mode specific speed, distance and emission information - if none of the corresponding input parameters were sampled from a distribution, call get_synthetic_from_trips to set synthetic trips and synthetic population - if none of the corresponding input parameters were sampled from a distribution, call get_all_distances.R to calculate distances