Run the set up script for ITHIM

Sets up the basic ITHIM object for onward calculation. Data loading, processing and harmonisation. Setting global values.

run_ithim_setup(
  seed = 1,
  CITY = "accra",
  speeds = NULL,
  PM_emission_inventory = NULL,
  CO2_emission_inventory = NULL,
  setup_call_summary_filename = "setup_call_summary.txt",
  DIST_CAT = c("0-6 km", "7-9 km", "10+ km"),
  AGE_RANGE = c(0, 150),
  ADD_WALK_TO_BUS_TRIPS = T,
  ADD_BUS_DRIVERS = T,
  ADD_TRUCK_DRIVERS = T,
  ADD_MOTORCYCLE_FLEET = F,
  TEST_WALK_SCENARIO = F,
  TEST_CYCLE_SCENARIO = F,
  MAX_MODE_SHARE_SCENARIO = F,
  REFERENCE_SCENARIO = "Baseline",
  PATH_TO_LOCAL_DATA = NULL,
  NSAMPLES = 1,
  BUS_WALK_TIME = 5,
  RAIL_WALK_TIME = 15,
  MMET_CYCLING = 4.63,
  MMET_WALKING = 2.53,
  PM_CONC_BASE = 50,
  PM_TRANS_SHARE = 0.225,
  PA_DOSE_RESPONSE_QUANTILE = F,
  AP_DOSE_RESPONSE_QUANTILE = F,
  BACKGROUND_PA_SCALAR = 1,
  BACKGROUND_PA_CONFIDENCE = 1,
  INJURY_REPORTING_RATE = 1,
  CHRONIC_DISEASE_SCALAR = 1,
  DAY_TO_WEEK_TRAVEL_SCALAR = 7,
  SIN_EXPONENT_SUM = 2,
  CASUALTY_EXPONENT_FRACTION = 0.5,
  BUS_TO_PASSENGER_RATIO = 0.022,
  TRUCK_TO_CAR_RATIO = 0.21,
  FLEET_TO_MOTORCYCLE_RATIO = 0.01,
  PM_EMISSION_INVENTORY_CONFIDENCE = 1,
  DISTANCE_SCALAR_CAR_TAXI = 1,
  DISTANCE_SCALAR_WALKING = 1,
  DISTANCE_SCALAR_PT = 1,
  DISTANCE_SCALAR_CYCLING = 1,
  DISTANCE_SCALAR_MOTORCYCLE = 1
)

Arguments

seed
random seed

CITY
name of the city, and name of the directory containing city data files

speeds
named list of mode speeds

PM_emission_inventory
named list of mode emissions

setup_call_summary_filename
name to write setup call summary to

DIST_CAT
vector string of distance categories in the form '0-6'. (The unit is assumed to be the same as in the trip set.)

AGE_RANGE
vector of minimum and maximum ages to include

ADD_WALK_TO_BUS_TRIPS
logic: whether or not to add short walks to all PT trips

ADD_BUS_DRIVERS
logic: whether or not to add bus drivers

ADD_TRUCK_DRIVERS
logic: whether or not to add truck drivers

ADD_MOTORCYCLE_FLEET
logic: whether or not to add motorcycle fleet

TEST_WALK_SCENARIO
logic: whether or not to run the walk scenario

TEST_CYCLE_SCENARIO
logic: whether or not to run the cycle scenario

MAX_MODE_SHARE_SCENARIO
logic: whether or not to run the max mode share scenario

REFERENCE_SCENARIO
which scenario forms the reference for the health comparison

PATH_TO_LOCAL_DATA
path to CITY directory, if not using package

NSAMPLES
constant integer: number of samples to take

BUS_WALK_TIME
lognormal parameter: duration of walk to Bus

RAIL_WALK_TIME
lognormal parameter: duration of walk to Rail

MMET_CYCLING
lognormal parameter: mMETs when cycling

MMET_WALKING
lognormal parameter: mMETs when walking

PM_CONC_BASE
lognormal parameter: background PM2.5 concentration

PM_TRANS_SHARE
beta parameter: fraction of background PM2.5 attributable to transport

PA_DOSE_RESPONSE_QUANTILE
logic: whether or not to sample from PA RR DR functions

AP_DOSE_RESPONSE_QUANTILE
logic: whether or not to sample from AP RR DR functions

BACKGROUND_PA_SCALAR
lognormal parameter: reporting scalar for PA

BACKGROUND_PA_CONFIDENCE
beta parameter: confidence in accuracy of PA survey

INJURY_REPORTING_RATE
lognormal parameter: rate of injury reporting

CHRONIC_DISEASE_SCALAR
lognormal parameter: scalar for background disease rates

DAY_TO_WEEK_TRAVEL_SCALAR
beta parameter: rate of scaling travel from one day to one week

SIN_EXPONENT_SUM
lognormal parameter: linearity of injuries with respect to two modes. SIN_EXPONENT_SUM=2 means no safety in numbers.

CASUALTY_EXPONENT_FRACTION
beta parameter: casualty contribution to SIN_EXPONENT_SUM

BUS_TO_PASSENGER_RATIO
beta parameter: number of buses per passenger

TRUCK_TO_CAR_RATIO
beta parameter: number of trucks per car

FLEET_TO_MOTORCYCLE_RATIO
beta parameter: fraction of total motorcycles that's fleet

PM_EMISSION_INVENTORY_CONFIDENCE
beta parameter: confidence in accuracy of emission inventory

DISTANCE_SCALAR_CAR_TAXI
lognormal parameter: scalar for car distance travelled

DISTANCE_SCALAR_WALKING
lognormal parameter: scalar for walking distance travelled

DISTANCE_SCALAR_PT
lognormal parameter: scalar for PT distance travelled

DISTANCE_SCALAR_CYCLING
lognormal parameter: scalar for cycling distance travelled

DISTANCE_SCALAR_MOTORCYCLE
lognormal parameter: scalar for motorcycle distance travelled

Value

ithim_object list of objects for onward use.

Details

Parameters have two options: to be set to a constant, and to be sampled from a prespecified distribution. Each parameter is given as an argument of length 1 or 2. If length 1, it's constant, and set to the global environment. If length 2, a distribution is defined and sampled from NSAMPLE times. There are some exceptions, listed above